feat: move ChannelFeedFilter and ChannelFeedViewModel to commons
Move the channel feed filter and view model to the commons module,
using IAccount and ICacheProvider interfaces. The filter uses
Channel (already in commons) and IAccount.isAcceptable().
The amethyst-side VM extends the commons version, passing LocalCache.
