+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/feeds/ChannelFeedFilter.kt
@@ -18,23 +18,28 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.chats.publicChannels.dal
+package com.vitorpamplona.amethyst.commons.ui.feeds
 
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import com.vitorpamplona.amethyst.commons.model.Channel
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.LocalCache
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.Note
 
-class ChannelFeedViewModel(
-    channel: Channel,
-    account: Account,
-) : com.vitorpamplona.amethyst.commons.viewmodels.ChannelFeedViewModel(channel, account, LocalCache) {
-    class Factory(
-        val channel: Channel,
-        val account: Account,
-    ) : ViewModelProvider.Factory {
-        @Suppress("UNCHECKED_CAST")
-        override fun <T : ViewModel> create(modelClass: Class<T>): T = ChannelFeedViewModel(channel, account) as T
-    }
+class ChannelFeedFilter(
+    val channel: Channel,
+    val account: IAccount,
+) : AdditiveFeedFilter<Note>(),
+    ChangesFlowFilter<Note> {
+    override fun feedKey() = channel
+
+    override fun changesFlow() = channel.changesFlow()
+
+    // returns the last Note of each user.
+    override fun feed(): List<Note> = sort(channel.notes.filterIntoSet { key, it -> account.isAcceptable(it) })
+
+    override fun applyFilter(newItems: Set<Note>): Set<Note> =
+        newItems
+            .filter { channel.notes.containsKey(it.idHex) && account.isAcceptable(it) }
+            .toSet()
+
+    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
 }
