+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/viewmodels/ChannelFeedViewModel.kt
@@ -18,31 +18,15 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.chats.publicChannels.dal
+package com.vitorpamplona.amethyst.commons.viewmodels
 
 import com.vitorpamplona.amethyst.commons.model.Channel
-import com.vitorpamplona.amethyst.model.Account
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.ui.dal.AdditiveFeedFilter
-import com.vitorpamplona.amethyst.ui.dal.ChangesFlowFilter
-import com.vitorpamplona.amethyst.ui.dal.DefaultFeedOrder
+import com.vitorpamplona.amethyst.commons.model.IAccount
+import com.vitorpamplona.amethyst.commons.model.cache.ICacheProvider
+import com.vitorpamplona.amethyst.commons.ui.feeds.ChannelFeedFilter
 
-class ChannelFeedFilter(
+open class ChannelFeedViewModel(
     val channel: Channel,
-    val account: Account,
-) : AdditiveFeedFilter<Note>(),
-    ChangesFlowFilter<Note> {
-    override fun feedKey() = channel
-
-    override fun changesFlow() = channel.changesFlow()
-
-    // returns the last Note of each user.
-    override fun feed(): List<Note> = sort(channel.notes.filterIntoSet { key, it -> account.isAcceptable(it) })
-
-    override fun applyFilter(newItems: Set<Note>): Set<Note> =
-        newItems
-            .filter { channel.notes.containsKey(it.idHex) && account.isAcceptable(it) }
-            .toSet()
-
-    override fun sort(items: Set<Note>): List<Note> = items.sortedWith(DefaultFeedOrder)
-}
+    val account: IAccount,
+    cacheProvider: ICacheProvider,
+) : ListChangeFeedViewModel(ChannelFeedFilter(channel, account), cacheProvider)
